fix scale/filename logic

+++ setupSunriseRun_gfs.py
@@ -79,7 +79,7 @@ def generate_mcrx_config(run_dir, snap_dir, filename, stub_name, galprops_data, 
 	#move npixels to .config file
 
 	if run_type == 'images':
-                npix=galprops_data['image_npix'][idx]
+                npix=np.int32(galprops_data['image_npix'][idx])
 		mf.write('npixels     {:8d}\n'.format(npix))
 	elif run_type == 'ifu':
 		mf.write('npixels     400\n')
